worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.35",
    "openai-agents>=0.7.0",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sendgrid>=6.12.5",
    "sqlalchemy>=2.0.46",
]
